+++ algorithm/faultinessreason.go
@@ -1,4 +1,4 @@
-package monitor
+package algorithm
 
 import (
 	"strconv"
