+++ validator/_config/config_1.yaml
@@ -1,7 +1,7 @@
 --- # config file for validator 1
 id: 1
 heightvoteset:
-  3: 
+  3:
     round: 3
     received_prevote:
       - type: PREVOTE
