+++ validator/_config/config_4.yaml
@@ -1,7 +1,7 @@
---- # config file for validator 2
-id: 2
+--- # config file for validator 4
+id: 4
 heightvoteset:
-  3: 
+  3:
     round: 3
     received_prevote:
       - type: PREVOTE
@@ -19,35 +19,21 @@ heightvoteset:
         round: 3
         value: 10
 
-      - type: PREVOTE
-        sender: 1
-        round: 3
-        value: 20
-
-      - type: PREVOTE
-        sender: 3
-        round: 3
-        value: 20
-      
-      - type: PREVOTE
-        sender: 4
-        round: 3
-        value: 20
-
     received_precommit: []
 
     sent_prevote:
       - type: PREVOTE
-        sender: 2
+        sender: 4
         round: 3
         value: 10
 
     sent_precommit:
       - type: PRECOMMIT
-        sender: 2
+        sender: 4
         round: 3
         value: 10
-  4: 
+
+  4:
     round: 4
     received_prevote:
       - type: PREVOTE
@@ -65,31 +51,18 @@ heightvoteset:
         round: 4
         value: 20
 
-    received_precommit:
-      - type: PRECOMMIT
-        sender: 2
-        round: 4
-        value: 20
-
-      - type: PRECOMMIT
-        sender: 3
-        round: 4
-        value: 20
-
-      - type: PRECOMMIT
-        sender: 4
-        round: 4
-        value: 20
+    received_precommit: []
 
     sent_prevote:
       - type: PREVOTE
-        sender: 2
+        sender: 4
         round: 4
         value: 20
 
     sent_precommit:
       - type: PRECOMMIT
-        sender: 2
+        sender: 4
         round: 4
         value: 20
+
